refactor: refactor reusable helpers
