+++ src/components/Agenda/AgendaForm.vue
@@ -268,7 +268,7 @@
 import debounce from 'lodash.debounce';
 import {
   daysDifference, formatDate, isToday, minutesDifference,
-} from '@/lib/date-fns';
+} from '@/common/helpers/date';
 import { AGENDA_CATEGORIES } from '@/static/data';
 import HeaderMenu from '@/common/components/HeaderMenu';
 import BaseButton from '@/common/components/BaseButton';
